requirements.txt
virny==0.6.0
autokeras==1.0.20
tensorflow==2.17.0
kmodes==0.12.2
numpy==1.23.5
matplotlib>=3.6.2
pandas>=1.5.2
altair>=4.2.0
scikit-learn>=1.2.0
lightgbm>=4.3.0
tqdm>=4.64.1
sklearn-utils>=0.0.15
seaborn>=0.12.1
fairlearn>=0.9.0
aif360>=0.6.1
BlackBoxAuditing>=0.1.54
impyute>=0.0.8
pymongo>=4.8.0
certifi>=2024.2.2
boto3>=1.34.155
fastapi>=0.112.0
uvicorn>=0.30.5
motor>=3.5.1
python-dotenv>=1.0.1